Instead of rewriting package json, simply install the proper webpack version as the last step.

+++ generators/root/index.js
@@ -8,16 +8,20 @@ module.exports = generator.Base.extend({
     this.argument('name', { type: String, required: true });
   },
 
+  /*
   conflicts: function configuring() {
     // Rewrite the webpack version the the last know working beta version
     // It is left here in plain sight, to remind about the pain updating to
     // a master branch without checking for the actual version.
     //
     // TODO: throw this out as soon as the next stable version is released
-    var config = JSON.parse(fs.readFileSync(this.destinationPath('package.json')));
-    config.devDependencies.webpack = '=2.1.0-beta.6';
-    fs.writeFileSync(this.destinationPath('package.json'), JSON.stringify(config));
+    if(fs.existsSync(this.destinationPath('package.json'))) {
+      var config = JSON.parse(fs.readFileSync(this.destinationPath('package.json')));
+      config.devDependencies.webpack = '=2.1.0-beta.6';
+      fs.writeFileSync(this.destinationPath('package.json'), JSON.stringify(config));
+    }
   },
+  */
 
   writing: function writing() {
     /* Some base functionality needs to be overwritten, so we force yeoman to do
@@ -68,5 +72,4 @@ module.exports = generator.Base.extend({
       this.destinationPath('.eslintrc')
     );
   }
-
 });
